Merged PR 5210: Removed /datasets from Frontend ARM Template datasetsClient:ApiEndpoint
Removed /datasets from Frontend ARM Template datasetsClient:ApiEndpoint at the request of Antony Gorman

I'm about to raise a PR that when deployed will need a change to the configured uri we have the datasets api client - currently we suffix the uri with /datasets (e.g. for local dev it is http://localhost:7013/api/datasets) - this now needs to be change to omit the /datasets portion when the ARM template sets it for the front end deployment (e.g. in local dev it will now be http://localhost:7013/api) - can you organise this? cheers

Related work items: #29738